feat: Move Ktor data sources and FcmPayloadParser to data module (Phase 21)
Move KtorNetworkDoorDataSource, KtorNetworkButtonDataSource,
KtorNetworkConfigDataSource, and FcmPayloadParser from androidApp to
the data module's commonMain source set. Replace java.net.URLDecoder
with a KMP-compatible percentDecode function. Add Ktor and kotlinx
serialization dependencies to the data module.
